observations: also move 5 sub-components + 2 naming fields fragments
The ObservationForm sub-components (Upload, Details, Specimen,
Projects, Lists) are only rendered from ObservationForm — nest them
under it as `Views::Controllers::Observations::Form::*`:

  Components::ObservationFormUpload  → Form::Upload
  Components::ObservationFormDetails → Form::Details
  Components::ObservationFormSpecimen → Form::Specimen
  Components::ObservationFormProjects → Form::Projects
  Components::ObservationFormLists   → Form::Lists

The naming fields are rendered from both the obs form *and* the
namings form — make them siblings under the namings cluster:

  Components::NamingFields         → Namings::Fields
  Components::NamingReasonsFields  → Namings::ReasonsFields

Convention: private-to-one-form sub-components nest inside that
form (`Form::Upload`); shared fragments live at the sibling
namespace level (`Namings::Fields`). Mirrors the Rails private-
helper / shared-partial distinction.

Inside the parent class bodies the references shorten via Ruby's
lexical lookup (`render(Upload.new(...))`, `render(Fields.new(...))`)
— cross-namespace callers use the full path
(`render(Views::Controllers::Observations::Namings::Fields.new(...))`).

371 tests pass (41 view + 330 controller). Lint clean.
